+++ src/test/converter2/issues/gh3006/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @document two words.md
+ * @module
+ */
+
+/**
+ * [link](./two%20words.md)
+ */
+export const x = 1;